requirements.txt
flask
flask_cors
googletrans
lxml
bs4
pathlib
igramscraper
python-firebase